+++ src/mako/application/cli/commands/app/ListRoutes.php
@@ -14,7 +14,7 @@ use mako\http\routing\Dispatcher;
 use mako\http\routing\Route;
 use mako\http\routing\Router;
 use mako\http\routing\Routes;
-use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\CommandArguments;
 use mako\reactor\attributes\CommandDescription;
 use mako\reactor\Command;
 use mako\utility\Arr;
@@ -31,7 +31,7 @@ use function str_pad;
  * Command that lists all registered routes.
  */
 #[CommandDescription('Lists all registered routes.')]
-#[Arguments(
+#[CommandArguments(
 	new Argument('-f|--filter', 'Filter routes using the route action, name or path', Argument::IS_OPTIONAL),
 	new Argument('-d|--detailed', 'Show more information about the route', Argument::IS_BOOL | Argument::IS_OPTIONAL),
 )]