+++ src/mako/application/cli/commands/cache/Clear.php
@@ -9,14 +9,14 @@ namespace mako\application\cli\commands\cache;
 
 use mako\cache\CacheManager;
 use mako\cli\input\arguments\Argument;
-use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\CommandArguments;
 use mako\reactor\attributes\CommandDescription;
 
 /**
  * Command that clears the cache.
  */
 #[CommandDescription('Clears the cache.')]
-#[Arguments(
+#[CommandArguments(
 	new Argument('-c|--configuration', 'Configuration name', Argument::IS_OPTIONAL),
 )]
 class Clear extends Command