+++ src/mako/application/cli/commands/migrations/Create.php
@@ -10,7 +10,7 @@ namespace mako\application\cli\commands\migrations;
 use mako\application\Application;
 use mako\cli\input\arguments\Argument;
 use mako\file\FileSystem;
-use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\CommandArguments;
 use mako\reactor\attributes\CommandDescription;
 use mako\reactor\Command;
 use Throwable;
@@ -23,7 +23,7 @@ use function vsprintf;
  * Command that creates a migration.
  */
 #[CommandDescription('Creates a new migration.')]
-#[Arguments(
+#[CommandArguments(
 	new Argument('-d|--description', 'Migration description', Argument::IS_OPTIONAL),
 	new Argument('-p|--package', 'Package name', Argument::IS_OPTIONAL),
 )]