+++ src/mako/application/cli/commands/migrations/Status.php
@@ -8,7 +8,7 @@
 namespace mako\application\cli\commands\migrations;
 
 use mako\cli\input\arguments\Argument;
-use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\CommandArguments;
 use mako\reactor\attributes\CommandDescription;
 
 use function count;
@@ -18,7 +18,7 @@ use function vsprintf;
  * Command that checks if there are any outstanding migrations.
  */
 #[CommandDescription('Checks if there are any outstanding migrations.')]
-#[Arguments(
+#[CommandArguments(
 	new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
 	new Argument('-e|--exit-code', 'Exits with 1 if there are outstanding migrations and 0 if there are none', Argument::IS_BOOL),
 )]