+++ src/mako/application/cli/commands/migrations/Up.php
@@ -8,14 +8,14 @@
 namespace mako\application\cli\commands\migrations;
 
 use mako\cli\input\arguments\Argument;
-use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\CommandArguments;
 use mako\reactor\attributes\CommandDescription;
 
 /**
  * Command that runs all outstanding migrations.
  */
 #[CommandDescription('Runs all outstanding migrations.')]
-#[Arguments(
+#[CommandArguments(
 	new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
 )]
 class Up extends Command